pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "numpy"]
build-backend = "setuptools.build_meta"

[project]
name = "magpiem"
version = "0.2.22"
authors = [
  { name="Frank Nightingale", email="fnight128@gmail.com" },
]
description = "Automated cleaning of sub-tomogram particle picking data"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["numpy",
"pandas",
"prettytable",
"scipy",
"plotly==5.15.0",
"pathlib",
"dash==2.15.0",
"dash_bootstrap_components==1.4.1",
"dash_bootstrap_templates==1.0.8",
"dash_extensions==1.0.1",
"dash_daq==0.5.0",
"flask==2.3.2",
"imodmodel",
"emfile",
"pyyaml",
"starfile",
"eulerangles"]

[project.scripts]
magpiem = "magpiem.dash.dash_ui:main"

[project.urls]
"Homepage" = "https://github.com/fnight128/MagpiEM"
"Bug Tracker" = "https://github.com/fnight128/MagpiEM/issues"